캘린더 반복일정 삭제 구현 중 ..1

+++ src/main/java/com/sjs/jsvill/dto/CalendarDTO.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class CalendarDTO {
     private Long calendarRowid;
     private Long groupRowid;
-    private Integer bundleId;
+    private Long bundleId;
     private String title;
     private String description;
     private String start;
